Rewrite code to pytorch-lightning. - much simpler code with better overview - more options - `DistributedDataParallel` !!!
Other important notes:
- Previously the model pretraining was WITHOUT the 'ab'-part of the image. I don't know how/why it worked before, but now the `Pretrainer` automatically adds a dummy 'ab'-part to the image (via `torch.zeros(...)`.
- TLDR: Major bug in pretraining fixed.

+++ src/utils/utils.py
@@ -62,7 +62,7 @@ class GANLoss(nn.Module):
     def get_labels(self, preds, target_is_real_img):
         labels = self.real_label if target_is_real_img \
             else self.fake_label
-        return labels.expand_as(preds)
+        return labels.type_as(preds).expand_as(preds)
 
     def __call__(self, preds, target_is_real_img):
         labels = self.get_labels(preds, target_is_real_img)
